+++ httpie/output/streams.py
@@ -1,8 +1,9 @@
 from abc import ABCMeta, abstractmethod
 from itertools import chain
-from typing import Callable, Iterable, Union
+from typing import Callable, Iterable, Tuple, Union
 
 from ..cli.constants import EMPTY_FORMAT_OPTION
+from ..codec import TextDecoderStrategy
 from ..context import Environment
 from ..constants import UTF8
 from ..models import HTTPMessage, HTTPResponse
@@ -137,15 +138,25 @@ class PrettyStream(EncodedStream):
         super().__init__(**kwargs)
         self.formatting = formatting
         self.conversion = conversion
-        self.mime = self.get_mime()
+        self.mime, self.charset = self._get_mime_and_charset()
 
-    def get_mime(self) -> str:
-        mime = parse_header_content_type(self.msg.content_type)[0]
-        if isinstance(self.msg, HTTPResponse):
-            forced_content_type = self.formatting.options['response']['as']
-            if forced_content_type != EMPTY_FORMAT_OPTION:
-                mime = parse_header_content_type(forced_content_type)[0] or mime
-        return mime
+    def _get_mime_and_charset(self) -> Tuple[str, str]:
+        # Defaults from the message `Content-Type`.
+        mime, options = parse_header_content_type(self.msg.content_type)
+        charset = options.get('charset') or ''
+
+        if not isinstance(self.msg, HTTPResponse):
+            return mime, charset
+
+        # The response `Content-Type` could be overridden from options.
+        forced_content_type = self.formatting.options['response']['as']
+        if forced_content_type == EMPTY_FORMAT_OPTION:
+            return mime, charset
+
+        forced_mime, forced_options = parse_header_content_type(forced_content_type)
+        mime = forced_mime or mime
+        charset = forced_options.get('charset') or ''
+        return mime, charset
 
     def get_headers(self) -> bytes:
         return self.formatting.format_headers(
@@ -206,5 +217,9 @@ class BufferedPrettyStream(PrettyStream):
 
         if converter:
             self.mime, body = converter.convert(body)
+
+        # Decode the body using the most appropriate encoding.
+        # This is a no-op if it is already a string (likely altered by a converter).
+        body = TextDecoderStrategy(self.charset).decode(body)
 
         yield self.process_body(body)